Added button in captionview
Made it more generalized, capable of handling any string null or not

Added caption button for order using flag

3. In the Caption view, add an additional button at the very bottom,
with width spanning the entire phone width that says "Add to Order".
This button is only visible when the menu is being used to place an
order (i.e. based on a flag set during the segue).